+++ NEWS.md
@@ -1,5 +1,9 @@
-# chevron 0.2.0.9000
+# chevron 0.2.0.9004
 
+* Update `LBT06` template to use `PARAM` as row split.
+* Convert `AVISIT` to factor according to `AVISITN` as order in preprocessing.
+* Add `CFBT01` template. `VST01`, `EGT01` and `LBT01` are now all following `CFBT01`. By default parameters are by page displayed.
+* Update `EGT03` to use `ACTARMCD` as default arm variable, and remove the preprocessing of filtering to "HR".
 * Specify minimal version of dependencies.
 
 # chevron 0.2.0